ci(event_source): Splitting tests for improved maintainability and readability
